Избавились от файла PE030
включен в состав projectEuler.cpp
